requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
exifread>=3.0.0
pymediainfo>=6.0.0
pyinstaller>=6.0.0
